requirements.txt
aiogram==2.24
python-dotenv==0.21.1
openai==0.26.4